Adds router setup for easy receiver injection
